Reorganized mydeployments components
