+++ frontend/dashboard-mp/src/components/MyDeployments/DeploymentsPage.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import { Navigate } from "react-router-dom";
 import { useAuth } from "react-oidc-context";
-import { Alert, Container } from "react-bootstrap";
-import { getDeployments, getSubscribedServices } from "../configs/marketplaceConfig";
+import { Alert, Container, Row, Col } from "react-bootstrap";
+import { getDeployments } from "../../configs/marketplaceConfig";
 import DeploymentsTable from "./DeploymentsTable";
 
 function DeploymentsPage (props) {
@@ -25,7 +25,7 @@ function DeploymentsPage (props) {
         } else {
             if (auth.isAuthenticated) {
                 // Get deployments from marketplace
-                getDeployments(auth.user.access_token).then((response) => {
+                getDeployments(auth.user?.access_token).then((response) => {
                     if (response.status === 200) {
                         response.json().then((data) => {
                             if (data.deployments === undefined) {
@@ -52,7 +52,7 @@ function DeploymentsPage (props) {
                 })
             }
         }
-    }, [auth.isLoading, setIsLoading])
+    }, [auth.isLoading, setIsLoading, auth.isAuthenticated, auth.user?.access_token])
 
     if (!auth.isLoading && shouldRedirect) {
         return <Navigate to="/" />;
@@ -63,10 +63,21 @@ function DeploymentsPage (props) {
         if (auth.isAuthenticated) {
             return (
                 <Container>
-                    <DeploymentsTable deployments={deployments} isLoading={isLoading} setIsLoading={setIsLoading} error={error} setError={setError}/>
-                    <Alert variant="danger" show={error !== ""}>
-                        {error}
-                    </Alert>
+                    <Row className="text-center m-3">
+                        <Col>
+                            <h1>My Deployments</h1>
+                        </Col>
+                    </Row>
+                    <Row className="m-2">
+                        <Col>
+                        <DeploymentsTable deployments={deployments} isLoading={isLoading} setIsLoading={setIsLoading} error={error} setError={setError}/>
+                        </Col>
+                    </Row>
+                    <Row>
+                        <Alert variant="danger" show={error !== ""}>
+                            {error}
+                        </Alert>
+                    </Row>
                 </Container>
             )
         } else {
